Move context/cloud => api/cloud
