+++ aci/backend.go
@@ -34,7 +34,7 @@ import (
 	"github.com/docker/compose-cli/api/volumes"
 
 	apicontext "github.com/docker/compose-cli/context"
-	"github.com/docker/compose-cli/context/cloud"
+	"github.com/docker/compose-cli/api/cloud"
 	"github.com/docker/compose-cli/context/store"
 )
