+++ ecs/local/context.go
@@ -19,7 +19,7 @@ package local
 import (
 	"context"
 
-	"github.com/docker/compose-cli/context/cloud"
+	"github.com/docker/compose-cli/api/cloud"
 	"github.com/docker/compose-cli/ecs"
 	"github.com/docker/compose-cli/errdefs"
 )
